+++ src/book-components/AvatarProfile/AvatarChip/index.ts
@@ -1,0 +1,3 @@
+// Barrel file for AvatarChip exports
+export * from './AvatarChip';
+export * from './AvatarChipFromSource';